refactor(worktree): move writeFeedbackWrapper to feedback-wrapper.ts
